# refactor: use gojson to marshal config for SHA