**What this PR does / why we need it**:

Use [`gojson`](https://github.com/goccy/go-json) to marshal config to JSON to get config SHA.

This speeds up the marshalling of config to JSON by about 35%, using roughly the same amount of memory and the same amount of memory allocations.

Results (run locally):

```
$ benchstat /tmp/old /tmp/new
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/deckgen
                      │  /tmp/old   │              /tmp/new               │
                      │   sec/op    │   sec/op     vs base                │
DeckgenGenerateSHA-10   31.32µ ± 0%   20.25µ ± 1%  -35.36% (p=0.000 n=10)

                      │   /tmp/old   │              /tmp/new               │
                      │     B/op     │     B/op      vs base               │
DeckgenGenerateSHA-10   10.95Ki ± 0%   10.87Ki ± 0%  -0.70% (p=0.000 n=10)

                      │  /tmp/old  │            /tmp/new            │
                      │ allocs/op  │ allocs/op   vs base            │
DeckgenGenerateSHA-10   12.00 ± 0%   12.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

Results can be observed at https://kong.github.io/kubernetes-ingress-controller/dev/bench/ for https://github.com/Kong/kubernetes-ingress-controller/commit/2501474b2cf5eeb4104531811a4b3396c89de588 or at https://github.com/Kong/kubernetes-ingress-controller/actions/runs/5377294841

**Special notes for your reviewer**:

<!-- Here you can add any open questions or notes that you might have for reviewers -->

**PR Readiness Checklist**:

Complete these before marking the PR as `ready to review`:

- [ ] the `CHANGELOG.md` release notes have been updated to reflect any significant (and particularly user-facing) changes introduced by this PR