@czeslavo 2023-06-26
Do you think it would be worth noting in the changelog? IMHO we could mention it in the "Under the hood" section, WDYT?