@rainest 2023-06-26
I'm ambivalent on a changelog entry since the SHA computation isn't something you usually think too much about. Can't think of objections--they say they're compatible with the stdlib version, so as long as they don't break that I wouldn't expect issues.

Should we not also use this elsewhere we're currently using `json.Marshal` and `json.Unmarshal`? The `internal/dataplane/sendconfig/inmemory.go` in particular is operating on a similarly large struct.