@pmalek 2023-06-26
> I'm ambivalent on a changelog entry since the SHA computation isn't something you usually think too much about. Can't think of objections--they say they're compatible with the stdlib version, so as long as they don't break that I wouldn't expect issues.


I've added the changelog entry as that's something that people may refer to in case we'll observe performance gains.

> 
> Should we not also use this elsewhere we're currently using `json.Marshal` and `json.Unmarshal`? The `internal/dataplane/sendconfig/inmemory.go` in particular is operating on a similarly large struct.

We should definitely consider this but first: measure! Add a benchmark, gather some data points (so that it's visible at: https://kong.github.io/kubernetes-ingress-controller/dev/bench/) and then propose a patch so that we can observe the change (and claim the bragging rights).